+++ lib/upgradedb/postgres.2019102100.php
@@ -23,11 +23,13 @@
 
 $this->BeginTrans();
 
-$this->Execute("
-    DELETE FROM rttickets WHERE id NOT IN (SELECT DISTINCT ticketid FROM rtmessages);
-    ALTER TABLE rttickets ADD COLUMN modtime integer NOT NULL DEFAULT 0;
-    UPDATE rttickets t SET modtime = (SELECT MAX(createtime) FROM rtmessages m WHERE m.ticketid = t.id)
-");
+if (!$this->ResourceExists('rttickets.modtime', LMSDB::RESOURCE_TYPE_COLUMN)) {
+    $this->Execute("
+        DELETE FROM rttickets WHERE id NOT IN (SELECT DISTINCT ticketid FROM rtmessages);
+        ALTER TABLE rttickets ADD COLUMN modtime integer NOT NULL DEFAULT 0;
+        UPDATE rttickets t SET modtime = (SELECT MAX(createtime) FROM rtmessages m WHERE m.ticketid = t.id)
+    ");
+}
 
 $this->Execute("UPDATE dbinfo SET keyvalue = ? WHERE keytype = ?", array('2019102100', 'dbversion'));
